feat: parse forward expressions from CLI
Parse multiple onionpipe-style expressions from command line args.
