+++ src/parse/addr.rs
@@ -1,0 +1,74 @@
+// Some parsing functions adapted from https://github.com/zupzup/rust-nom-parsing, which is
+// licensed:
+//
+// Copyright 2020 Mario Zupan <mario@zupzup.org>
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use nom::{
+    bytes::complete::tag,
+    character::complete::one_of,
+    error::{context, VerboseError},
+    multi::{count, many_m_n},
+    sequence::{terminated, tuple},
+    Err as NomErr, IResult,
+};
+
+pub type Res<T, U> = IResult<T, U, VerboseError<T>>;
+
+#[derive(Debug, PartialEq, Eq)]
+pub enum Host {
+    HOST(String),
+    IP4([u8; 4]),
+}
+
+pub fn ip(input: &str) -> Res<&str, Host> {
+    context(
+        "ip",
+        tuple((count(terminated(ip_num, tag(".")), 3), ip_num)),
+    )(input)
+    .map(|(next_input, res)| {
+        let mut result: [u8; 4] = [0, 0, 0, 0];
+        res.0
+            .into_iter()
+            .enumerate()
+            .for_each(|(i, v)| result[i] = v);
+        result[3] = res.1;
+        (next_input, Host::IP4(result))
+    })
+}
+
+pub fn ip_num(input: &str) -> Res<&str, u8> {
+    context("ip number", n_to_m_digits(1, 3))(input).and_then(|(next_input, result)| {
+        match result.parse::<u8>() {
+            Ok(n) => Ok((next_input, n)),
+            Err(_) => Err(NomErr::Error(VerboseError { errors: vec![] })),
+        }
+    })
+}
+
+pub fn port(input: &str) -> Res<&str, u16> {
+    context("port", n_to_m_digits(2, 5))(input).and_then(|(next_input, res)| {
+        match res.parse::<u16>() {
+            Ok(n) => Ok((next_input, n)),
+            Err(_) => Err(NomErr::Error(VerboseError { errors: vec![] })),
+        }
+    })
+}
+
+fn n_to_m_digits<'a>(n: usize, m: usize) -> impl FnMut(&'a str) -> Res<&str, String> {
+    move |input| {
+        many_m_n(n, m, one_of("0123456789"))(input)
+            .map(|(next_input, result)| (next_input, result.into_iter().collect()))
+    }
+}
